cache: remove no longer needed get_additional_housenumbers_html()
Because it used write_additional_housenumbers() internally, which
already uses the JSON cache. This way the locale-dependent HTML cache
can be removed.
