+++ src/i18n/tests.rs
@@ -13,23 +13,31 @@
 use super::*;
 
 /// Context manager for translate().
-struct LanguageContext<'a> {
-    ctx: &'a context::Context,
-}
+struct LanguageContext {}
 
-impl<'a> LanguageContext<'a> {
+impl LanguageContext {
     /// Switches to the new language.
-    fn new(ctx: &'a context::Context, language: &str) -> Self {
+    fn new(ctx: &context::Context, language: &str) -> Self {
         set_language(ctx, language);
-        LanguageContext { ctx }
+        LanguageContext {}
     }
 }
 
-impl<'a> Drop for LanguageContext<'a> {
+impl<'a> Drop for LanguageContext {
     /// Switches back to the old language.
     fn drop(&mut self) {
-        set_language(self.ctx, "en");
+        reset_language();
     }
+}
+
+/// Resets the language.
+pub fn reset_language() {
+    LANGUAGE.with(|it| {
+        *it.borrow_mut() = None;
+    });
+    TRANSLATIONS.with(|it| {
+        *it.borrow_mut() = None;
+    });
 }
 
 /// Tests translate().
@@ -38,4 +46,12 @@ fn test_translate() {
     let ctx = context::tests::make_test_context().unwrap();
     let _lc = LanguageContext::new(&ctx, "hu");
     assert_eq!(translate("Area"), "Terület");
+}
+
+/// Tests get_language() when its value is None.
+#[test]
+fn test_get_language_none() {
+    reset_language();
+
+    assert_eq!(get_language(), "en");
 }
